v0.2 - Failed to implement Axios.
Can't seem to get proper key-value pairs after successful logins. I'd like to get a key-value pair for each parameter (id, password, name, etc), but this program is stuck with a single key with an extensive array (that contains all of the needed parameters) for a value. Makes the parameters individually inaccessible, at least with current implementations.